Define meta page more broadly so it can be used by module metas or final metas, but only by consent
